+++ crates/apps/src/lib/node/ledger/shell/governance.rs
@@ -2,6 +2,7 @@ use namada::core::collections::HashMap;
 use namada::core::encode;
 use namada::core::event::EmitEvents;
 use namada::core::storage::Epoch;
+use namada::governance::event::ProposalEvent;
 use namada::governance::pgf::storage::keys as pgf_storage;
 use namada::governance::pgf::storage::steward::StewardDetail;
 use namada::governance::pgf::{storage as pgf, ADDRESS};
@@ -17,7 +18,6 @@ use namada::governance::{
 };
 use namada::ibc;
 use namada::ledger::events::extend::{ComposeEvent, Height};
-use namada::ledger::governance::utils::ProposalEvent;
 use namada::proof_of_stake::bond_amount;
 use namada::proof_of_stake::parameters::PosParams;
 use namada::proof_of_stake::storage::{
